Another fix

+++ spring_backend/src/main/java/com/capstoneproject/POJOS/controller/MainController.java
@@ -127,7 +127,9 @@ public class MainController
             homes.forEach(home ->
             {
                 Optional<HomePolicy> homePolicy = homePolicyRepository.getByHomeId(home.getId());
-                homePolicyRepository.delete(homePolicy.get());
+                if(homePolicy.isPresent()){
+                    homePolicyRepository.delete(homePolicy.get());
+                }
             });
             homeRepository.deleteAll(homes);
 
@@ -136,7 +138,9 @@ public class MainController
             autos.forEach(auto ->
             {
                 Optional<AutoPolicy> autoPolicy = autoPolicyRepository.getByVehicleId(auto.getId());
-                autoPolicyRepository.delete(autoPolicy.get());
+                if(autoPolicy.isPresent()){
+                    autoPolicyRepository.delete(autoPolicy.get());
+                }
             });
             autoRepository.deleteAll(autos);
 
@@ -566,7 +570,7 @@ public class MainController
     //Quote REST
 
     /**
-     * Home Quote
+     * Home Quote - Does not write to database. Option to access route should only be available to registered Home views.
      *
      * @param id
      * @param home_id
@@ -589,6 +593,12 @@ public class MainController
         return homeQuote;
     }
 
+    /**
+     * Auto Quote - Does not write to database. Option to access route should only be available to registered vehicle views.
+     * @param id
+     * @param auto_id
+     * @return AutoQuote
+     */
     @CrossOrigin
     @GetMapping(path = RESTNouns.USER + RESTNouns.USER_ID + RESTNouns.QUOTE + RESTNouns.AUTO + RESTNouns.AUTO_ID)
     public @ResponseBody Optional<AutoQuote> getAutoQuote(@PathVariable Integer id, @PathVariable Integer auto_id)
@@ -609,6 +619,13 @@ public class MainController
 
     // HomePolicy REST
 
+    /**
+     * Save a Home Policy to Database
+     * @param id
+     * @param home_id
+     * @param startDate
+     * @return
+     */
     @CrossOrigin
     @PostMapping(path = RESTNouns.USER + RESTNouns.USER_ID + RESTNouns.POLICY + RESTNouns.HOME + RESTNouns.HOME_ID)
     public @ResponseBody String addNewHomePolicy(@PathVariable Integer id,
@@ -624,19 +641,29 @@ public class MainController
             if (home.isPresent())
             {
                 homePolicy = Optional.of(PolicyBuilder.getNewHomePolicy(startDate, startDate.plusDays(365), home.get(), user.get()));
+                homePolicyRepository.save(homePolicy.get());
+                return "Home Policy Created!";
             }
-            homePolicyRepository.save(homePolicy.get());
-            return "Home Policy Created!";
+            else{
+                return "Home not registered";
+            }
         }
         else
         {
-            return "Please Register or Login!";
+            return "Please Register User";
         }
 
     }
 
     //AutoPolicy REST
 
+    /**
+     * Save an Auto Policy to Database
+     * @param id
+     * @param auto_id
+     * @param startDate
+     * @return
+     */
     @CrossOrigin
     @PostMapping(path = RESTNouns.USER + RESTNouns.USER_ID + RESTNouns.POLICY + RESTNouns.AUTO + RESTNouns.AUTO_ID)
     public @ResponseBody String addNewAutoPolicy(@PathVariable Integer id,
@@ -651,16 +678,25 @@ public class MainController
 
             Optional<Vehicle> auto = autoRepository.findById(auto_id);
             Optional<Driver> driver = driverRepository.findByUserId(id);
-            if (auto.isPresent())
-            {
-                autoPolicy = Optional.of(PolicyBuilder.getNewAutoPolicy(startDate, startDate.plusDays(365), auto.get(), driver.get(), user.get()));
+            if(driver.isPresent()){
+                if (auto.isPresent())
+                {
+                    autoPolicy = Optional.of(PolicyBuilder.getNewAutoPolicy(startDate, startDate.plusDays(365), auto.get(), driver.get(), user.get()));
+                    autoPolicyRepository.save(autoPolicy.get());
+                    return "Auto Policy Created!";
+                }
+                else
+                {
+                    return "Vehicle not Found";
+                }
             }
-            autoPolicyRepository.save(autoPolicy.get());
-            return "Auto Policy Created!";
+            else{
+                return "Please register as a Driver";
+            }
         }
         else
         {
-            return "Please register or login!";
+            return "Please register User.";
         }
 
     }
